Implement addAttachmentToCard function
Following the style of the code so far, I based this on the existing
addMemberToCard function.  This lets you add an attachment to a card by
specifying a URL starting with http:// or https:// and is the simplest
way to add attachments in Trello.

It also includes tests: Following the style of the tests so far, this
tests that the provided card ID and attachment URL are correctly
parametrised into the POST request.

This resolves #2.