Expect makeRequest's fn to complete only once
An HTTP request only completes once, therefore it is sufficient to
attach a one-time listener .once() to fn, which will be removed once
triggered and does not count towards the max listeners count like .on().
This solves #5, albeit in a less drastic manner than PR #10.

The tests have also been updated to fire "once: complete" events instead
of "on: complete". However, in practice, older versions of the restler
library did not inherit properly from node's Event, and would fire
'complete' more than once, therefore also breaking the one-time
listener, so this commit depends on issue #9 (for an updated restler
version) to work correctly.