Add copyBoard functionality
This is not a separate resource in the API.  Rather, if you include a
source board ID in the post request for creating a board (same as used
by the addBoard function) then that board will be used as a base to copy
for the new one.  Because its parameters have very little overlap with
the existing addBoard function it seemed more semantic to create a new
function for copying boards.

This resolves #80.

API docs: https://developers.trello.com/reference#boardsid